// web/src/api/models/createPersonRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface CreatePersonRequest {
    /**
     * Display name of the person
     * @minLength 1
     * @maxLength 100
     */
    displayName?: string;
    /** Birth date of the person */
    birthDate?: string;
    /** Gender of the person */
    gender?: string;
}