// web/src/api/models/createRelationshipRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type {CreateRelationshipRequestStatus} from './createRelationshipRequestStatus';
import type {CreateRelationshipRequestType} from './createRelationshipRequestType';

export interface CreateRelationshipRequest {
    personAId: string;
    personBId: string;
    type: CreateRelationshipRequestType;
    status?: CreateRelationshipRequestStatus;
    biological?: boolean;
    startDate?: string;
    endDate?: string;
}