// web/src/api/models/relationshipResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type {PersonResponse} from './personResponse';
import type {RelationshipResponseStatus} from './relationshipResponseStatus';
import type {RelationshipResponseType} from './relationshipResponseType';

export interface RelationshipResponse {
    id?: string;
    personA?: PersonResponse;
    personB?: PersonResponse;
    type?: RelationshipResponseType;
    status?: RelationshipResponseStatus;
    biological?: boolean;
    startDate?: string;
    endDate?: string;
}